Move client creation integ tests to functional
